ref: invalid meeting date exception

+++ src/main/java/com/github/krolikow/meetingsplanner/meetings/MeetingServiceImpl.java
@@ -1,9 +1,11 @@
 package com.github.krolikow.meetingsplanner.meetings;
 
+import com.github.krolikow.meetingsplanner.common.exceptions.InvalidMeetingDateException;
 import com.github.krolikow.meetingsplanner.common.exceptions.MeetingNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -35,11 +37,24 @@ public class MeetingServiceImpl implements MeetingService {
 
     @Override
     public Meeting save(Meeting meeting) {
+        if (meeting.getStartDate().isAfter(meeting.getEndDate())) {
+            throw new InvalidMeetingDateException("Start date cannot be after end date.");
+        }
+        if (meeting.getStartDate().isEqual(meeting.getEndDate())) {
+            throw new InvalidMeetingDateException("Start date cannot be equal end date.");
+        }
+        if (meeting.getStartDate().isBefore(LocalDateTime.now())) {
+            throw new InvalidMeetingDateException("Start date cannot be in the past.");
+        }
         return meetingRepository.save(meeting);
     }
 
     @Override
     public void deleteById(int id) {
+        Meeting meeting = findById(id);
+        if (meeting == null) {
+            throw new MeetingNotFoundException(id);
+        }
         meetingRepository.deleteById(id);
     }
 }
